add session_username cookie
Since there is no server-side rendering and session cookie is set with
attribute HttpOnly, client must have some other way to
check if they are logged in. So it's either creating some endpoint
specifically for this, or just sending a cookie with name back
(which I have implemented in this commit).

Refer to https://stackoverflow.com/questions/60826884/jwt-served-via-httponly-cookie-with-someway-to-find-out-is-logged-in/60841977#60841977.